refactor: split bpf tracer to a separate mod
Move the large chunk of bpf tracer code to a separate mod.
Use builder pattern.
